GRD-40393: Added --restart-mongodb flag to mongodb conf script

+++ mongodb-server-configuration/linux/configureMongodb.sh
@@ -5,21 +5,32 @@ dest=$1
 format=$2
 path=$3
 filter=$4
-logfile=$5
-#printf "%s: Params passed to mongod script:\nDESTINATION=%s\n\tFORMAT=%s\n\tPATH=%s\n\tFILTER=%s\n\tLOG=%s\n" $(date +"%Y-%m-%dT%H:%M:%SZ") $dest $format $path "$filter" $logfile|& tee -a $logfile
+restart_mongodb=$5
+logfile=$6
+#printf "%s: Params passed to mongod script:\nDESTINATION=%s\n\tFORMAT=%s\n\tPATH=%s\n\tFILTER=%s\n\tRESTART_MONGODB=%s\n\tLOG=%s\n" $(date +"%Y-%m-%dT%H:%M:%SZ") $dest $format $path "$filter" $restart_mongodb $logfile|& tee -a $logfile
 
 #Find mongod pid if exists
 mon_pid=$(service mongod status|grep -Eo '[0-9]+')
 printf "%s: Mongod pid: %s.\n" $(date +"%Y-%m-%dT%H:%M:%SZ") $mon_pid|& tee -a $logfile
 
-#Find mongod configuration file path
-mongod_conf=$( readlink  -f /*/mongod.conf )
-printf "%s: Mongod configuration file path is: %s.\n" $(date +"%Y-%m-%dT%H:%M:%SZ") $mongod_conf|& tee -a $logfile
+#Find mongod configuration file path to edit
+mongod_conf_path=$( readlink  -f /*/mongod.conf )
+printf "%s: Mongod configuration file path is: %s.\n" $(date +"%Y-%m-%dT%H:%M:%SZ") $mongod_conf_path|& tee -a $logfile
 
-#Backup configuration file
-#current_time=$(date "+%Y%m%d-%H%M%S")
-#cp $mongod_conf $mongod_conf.$current_time
-#echo "created $mongod_conf.$current_time as a backup configuration file" |& tee -a $logfile
+current_time=$(date "+%Y%m%d-%H%M%S")
+if [ "$restart_mongodb" = true ];
+then
+    #Backup configuration file
+    cp $mongod_conf_path $mongod_conf_path.$current_time
+    mongod_conf=$mongod_conf_path
+    echo "created $mongod_conf.$current_time as a backup configuration file" |& tee -a $logfile
+else
+    #Edit example file
+    cp $mongod_conf_path $mongod_conf_path.guardium_example.$current_time
+    mongod_conf=$mongod_conf_path.guardium_example.$current_time
+    echo "created $mongod_conf.guardium_example.$current_time as an example" |& tee -a $logfile
+    echo $mongod_conf
+fi
 
 #Find mongod pid file path configuration
 printf "%s: Mongod pid file path is:%s.\n" $(date +"%Y-%m-%dT%H:%M:%SZ") $(grep -o 'pidFilePath[^#]*' $mongod_conf | cut -d ':' -f 2)|& tee -a $logfile
@@ -33,7 +44,7 @@ then
     endRange=$((startRange+10))
     ignoreCommentsString="[^#]*//1;x;s/#.*//"
     sed -i -r "$startRange,$endRange{h;s/$ignoreCommentsString;s/destination:.*/destination: $dest/g;G;s/(.*)\n/\1/}" $mongod_conf
-    sed -i -r "$startRange,$endRange{h;s/$ignoreCommentsString;s/filter:.*/filter: $filter/g;G;s/(.*)\n/\1/}" $mongod_conf
+    sed -i -r "$startRange,$endRange{h;s/$ignoreCommentsString;s/filter:.*/filter: \'$filter\'/g;G;s/(.*)\n/\1/}" $mongod_conf
     sed -i "s/#setParameter: {auditAuthorizationSuccess: true}/setParameter: {auditAuthorizationSuccess: true}/g" $mongod_conf
     if [ "$dest" = "file" ];
     then
@@ -67,8 +78,13 @@ fi
 
 
 #Restart mongod service after configuration has changed
-service mongod stop
-rm -f /tmp/mongodb-27017.sock
-sleep 2
-service mongod start
-printf "%s: Restarted mongod.\n" $(date +"%Y-%m-%dT%H:%M:%SZ") |& tee -a $logfile
+if [ "$restart_mongodb"=true ];
+then
+    service mongod stop
+    rm -f /tmp/mongodb-27017.sock
+    sleep 2
+    service mongod start
+    printf "%s: Restarted mongod.\n" $(date +"%Y-%m-%dT%H:%M:%SZ") |& tee -a $logfile
+else
+    printf "%s: Changed were made in example file. Mongod was not restarted.\n" $(date +"%Y-%m-%dT%H:%M:%SZ") |& tee -a $logfile
+fi
